Consolidate `get_entity` into `tests.common`
