+++ tests/test_cover.py
@@ -7,7 +7,6 @@ from collections.abc import Awaitable, Callable
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
-from slugify import slugify
 from zigpy.device import Device as ZigpyDevice
 import zigpy.profiles.zha
 import zigpy.types
@@ -16,6 +15,7 @@ import zigpy.zcl.foundation as zcl_f
 
 from tests.common import (
     find_entity_id,
+    get_entity,
     make_zcl_header,
     send_attributes_report,
     update_attribute_cache,
@@ -24,7 +24,6 @@ from tests.conftest import SIG_EP_INPUT, SIG_EP_OUTPUT, SIG_EP_PROFILE, SIG_EP_T
 from zha.application import Platform
 from zha.application.const import ATTR_COMMAND
 from zha.application.gateway import Gateway
-from zha.application.platforms import PlatformEntity
 from zha.application.platforms.cover import (
     ATTR_CURRENT_POSITION,
     STATE_CLOSED,
@@ -117,13 +116,7 @@ WCT = closures.WindowCovering.WindowCoveringType
 WCCS = closures.WindowCovering.ConfigStatus
 
 
-def get_entity(zha_dev: Device, entity_id: str) -> PlatformEntity:
-    """Get entity."""
-    entities = {
-        entity.PLATFORM + "." + slugify(entity.name, separator="_"): entity
-        for entity in zha_dev.platform_entities.values()
-    }
-    return entities[entity_id]
+
 
 
 async def test_cover_non_tilt_initial_state(  # pylint: disable=unused-argument
